refactor(language): move CapabilityResolver out of compiler.rs
CapabilityResolver, ReferenceClass, PrimaryReference,
DEFAULT_NODE_KINDS, bind_capabilities, and bind_capabilities_with_registry
move from compiler.rs into a new capability_resolver.rs sibling of
resolver.rs, which already shares this binding policy (Resolver wraps
a CapabilityResolver internally) — the binding-policy triplication the
audit flagged is easier to see and keep in sync when the type lives
next to its other consumer instead of buried in the compiler. Public
API path updated at the crate root (lib.rs re-exports from the new
module); internal callers (resolver.rs, registry/capability, tests,
examples) updated to the new path. No behavior change.
