+++ tests/language_pipeline.rs
@@ -8,9 +8,8 @@
 use std::sync::Arc;
 
 use tinyagents::graph::{END, NodeFuture};
-use tinyagents::language::compiler::{
-    BoxedNode, CapabilityResolver, NodeFactory, bind_capabilities, build_graph, compile,
-};
+use tinyagents::language::capability_resolver::{CapabilityResolver, bind_capabilities};
+use tinyagents::language::compiler::{BoxedNode, NodeFactory, build_graph, compile};
 use tinyagents::language::parser::parse_str;
 use tinyagents::language::types::{END as LANG_END, NodeSpec, Routing};
 use tinyagents::{Command, NodeContext, NodeResult, Result, TinyAgentsError};
